①simp协议头中加入result字段，用于response表明请求成功or失败 ②msgpack版本回退至官方的1.4版本 ③console中首先创建主协程，可以让应用层在init初始化时即可使用协程操作 ④APP_PRT支持新的SIMP协议，且在创建连接设置连接对应的协议类型，从而在连接回调时根据类型找到正确的操作对象 ⑤库级的simpack_server与registry交互的名字宣告部分完成 ⑥seria组件支持use_simp字段时 fix bug ⑦seria反序列化时生成的键值对kv返回原始内存的引用，不再使用拷贝加快速度 ⑧simple_simps--一个简单的使用simpack_server组件与域内其他服务进行通信的例子

+++ include/kbase/serialize.h
@@ -32,6 +32,6 @@ namespace crx
         int get_sharding_size(const char *data, size_t len);
 
         //反序列化操作，将数据流反序列化为map表
-        std::unordered_map<std::string, mem_ref> dump(const char *data, int len);
+        std::unordered_map<std::string, mem_ref> dump(const char *data, size_t len);
     };
 }
